+++ sources/tifffile/large_image_source_tifffile/__init__.py
@@ -240,10 +240,11 @@ class TifffileFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
         Find associated images from unused pages and series.
         """
         pagesInSeries = [p for s in self._tf.series for ll in s.pages.levels for p in ll.pages]
+        hashes = [p.hash for p in pagesInSeries if p.keyframe is not None]
         self._associatedImages = {}
         for p in self._tf.pages:
             if (p not in pagesInSeries and p.keyframe is not None and
-                    not len(set(p.axes) - set('YXS'))):
+                    p.hash not in hashes and not len(set(p.axes) - set('YXS'))):
                 id = 'image_%s' % p.index
                 entry = {'page': p.index}
                 entry['width'] = p.shape[p.axes.index('X')]